+++ src/App.jsx
@@ -4,6 +4,7 @@ import ProductDetail from './components/Products/ProductDetail/ProductDetail';
 import Cart from './pages/Cart/Cart';
 import Home from './pages/Home/Home';
 import Login from './pages/Login/Login';
+import Product from './pages/Product/Product';
 
 function App() {
   return (
@@ -11,7 +12,7 @@ function App() {
       <Routes>
         <Route path='/' element={<Layout />}>
           <Route index element={<Home />} />
-          <Route path='/product/:id' element={<ProductDetail />} />
+          <Route path='/product/:id' element={<Product />} />
           <Route path='/cart' element={<Cart />} />
         </Route>
         <Route path='/login' element={<Login />} />
